refactor(core): split parser.rs into parser.rs + parser_elements.rs
Extract element extraction helpers (~946 LOC) from parser.rs into the
new sibling module parser_elements.rs:

- extract_elements, extract_calls, extract_imports, extract_impl_methods
- extract_references, extract_impl_traits_from_tree, extract_def_use
- enclosing_function_name
- extract_imports_from_node, extract_import_item_name, collect_import_items
- extract_python_import_from

parser.rs drops from 1949 to 1038 LOC (47% reduction).
ElementExtractor and SemanticExtractor remain public at their existing paths.
All 237 tests pass without modification.

Closes #1294
